+++ src/cli/directories.rs
@@ -17,50 +17,59 @@ use row::MultiProgressContainer;
 #[derive(Args, Debug)]
 #[allow(clippy::struct_excessive_bools)]
 pub struct Arguments {
-    /// Select the action to scan (defaults to all).
-    action: String,
-
     /// Select directories to summarize (defaults to all). Use 'show directories -' to read from stdin.
     directories: Vec<PathBuf>,
 
+    /// Select directories that are included by the provided action.
+    #[arg(long, short, display_order = 0)]
+    action: Option<String>,
+
     /// Hide the table header.
-    #[arg(long, display_order = 0)]
+    #[arg(long, display_order = 0, requires = "action")]
     no_header: bool,
 
     /// Do not separate groups with newlines.
-    #[arg(long, display_order = 0)]
+    #[arg(long, display_order = 0, requires = "action")]
     no_separate_groups: bool,
 
     /// Show an element of each directory's value (repeat to show multiple elements).
-    #[arg(long, value_name = "JSON POINTER", display_order = 0)]
+    #[arg(
+        long,
+        value_name = "JSON POINTER",
+        display_order = 0,
+        requires = "action"
+    )]
     value: Vec<String>,
 
     /// Limit the number of groups displayed.
-    #[arg(short, long, display_order = 0)]
+    #[arg(short, long, display_order = 0, requires = "action")]
     n_groups: Option<usize>,
 
     /// Show completed directories.
-    #[arg(long, display_order = 0)]
+    #[arg(long, display_order = 0, requires = "action")]
     completed: bool,
 
     /// Show submitted directories.
-    #[arg(long, display_order = 0)]
+    #[arg(long, display_order = 0, requires = "action")]
     submitted: bool,
 
     /// Show eligible directories.
-    #[arg(long, display_order = 0)]
+    #[arg(long, display_order = 0, requires = "action")]
     eligible: bool,
 
     /// Show waiting directories.
-    #[arg(long, display_order = 0)]
+    #[arg(long, display_order = 0, requires = "action")]
     waiting: bool,
+
+    /// Show only directory names.
+    #[arg(long, default_value_t = false, display_order = 0, requires = "action")]
+    short: bool,
 }
 
 /// Show directories that match an action.
 ///
 /// Print a human-readable list of directories, their status, job ID, and value(s).
 ///
-#[allow(clippy::too_many_lines)]
 pub fn directories<W: Write>(
     options: &GlobalOptions,
     args: Arguments,
@@ -68,7 +77,20 @@ pub fn directories<W: Write>(
     output: &mut W,
 ) -> Result<(), Box<dyn Error>> {
     debug!("Showing directories.");
+    match &args.action {
+        Some(action) => print_matching(&action.clone(), options, args, multi_progress, output),
+        None => print_all(options, args, multi_progress, output),
+    }
+}
 
+#[allow(clippy::too_many_lines)]
+pub fn print_matching<W: Write>(
+    action_name: &str,
+    options: &GlobalOptions,
+    args: Arguments,
+    multi_progress: &mut MultiProgressContainer,
+    output: &mut W,
+) -> Result<(), Box<dyn Error>> {
     // Show directories with selected statuses.
     let mut show_completed = args.completed;
     let mut show_submitted = args.submitted;
@@ -88,10 +110,10 @@ pub fn directories<W: Write>(
 
     project
         .workflow()
-        .action_by_name(&args.action)
-        .ok_or_else(|| row::Error::ActionNotFound(args.action.clone()))?;
+        .action_by_name(action_name)
+        .ok_or_else(|| row::Error::ActionNotFound(action_name.to_string()))?;
 
-    let mut table = Table::new().with_hide_header(args.no_header);
+    let mut table = Table::new().with_hide_header(if args.short { true } else { args.no_header });
     table.header = vec![
         Item::new("Directory".to_string(), Style::new().underlined()),
         Item::new("Status".to_string(), Style::new().underlined()),
@@ -108,7 +130,7 @@ pub fn directories<W: Write>(
     }
 
     for action in &project.workflow().action {
-        if action.name() != args.action {
+        if action.name() != action_name {
             continue;
         }
 
@@ -166,6 +188,12 @@ pub fn directories<W: Write>(
                     Style::new().bold(),
                 ));
 
+                // Only show directory names when user requests short output.
+                if args.short {
+                    table.rows.push(Row::Items(row));
+                    continue;
+                }
+
                 // Status
                 row.push(status);
 
@@ -201,18 +229,48 @@ pub fn directories<W: Write>(
                 table.rows.push(Row::Items(row));
             }
 
-            if !args.no_separate_groups && group_idx != groups.len() - 1 {
+            if !args.no_separate_groups && group_idx != groups.len() - 1 && !args.short {
                 table.rows.push(Row::Separator);
             }
         }
 
-        table.rows.push(Row::Separator);
+        if !args.short {
+            table.rows.push(Row::Separator);
+        }
     }
 
     table.write(output)?;
     output.flush()?;
 
     project.close(multi_progress)?;
+
+    Ok(())
+}
+
+pub fn print_all<W: Write>(
+    options: &GlobalOptions,
+    args: Arguments,
+    multi_progress: &mut MultiProgressContainer,
+    output: &mut W,
+) -> Result<(), Box<dyn Error>> {
+    let project = Project::open(options.io_threads, &options.cluster, multi_progress)?;
+
+    let all_directories = project.state().list_directories();
+    let mut query_directories =
+        cli::parse_directories(args.directories, || Ok(all_directories.clone()))?;
+    query_directories.sort_unstable();
+    let all_directories = HashSet::<PathBuf>::from_iter(all_directories);
+
+    for directory in &query_directories {
+        if !all_directories.contains(directory) {
+            warn!(
+                "Directory '{}' not found in workspace.",
+                directory.display()
+            );
+            continue;
+        }
+        writeln!(output, "{}", directory.display())?;
+    }
 
     Ok(())
 }